pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "labelprinter"
version = "0.1.0"
description = "Brother VC-500W label printer control"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pillow>=8.0.0",
]

[project.optional-dependencies]
cups = [
    "pycups>=2.0.0",
]

[project.scripts]
label-text = "labelprinter.print_text:main"
label-direct = "labelprinter.__main__:main"
label-queue = "labelprinter.queue_manager:main"
label-queue-worker = "labelprinter.queue_worker:main"
label-queue-setup = "labelprinter.queue_setup:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["labelprinter*"]